python_backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
statsmodels>=0.13.0
yfinance>=0.2.0
requests>=2.25.0
python-dotenv>=0.19.0
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0 